Updated corpus_functions.R and topic_modeling_utilities.R w/ tibbles
This is more for the end-user as all we do is check if `tibble` is installed and if it is we use that for output instead of base-R. So internally `data.frame`s are still used.

Initial tests of in-place replacements of `data.frame` to `tibble` fails several unit-tests; this approach retains
backwards-compatability, and those who do not use `tibble` are also unaffected.